Refactored Urls for smooter dev experience, needs realignment in the entities endpoints, UNFINISHED
