+++ .env.compose
@@ -1,0 +1,25 @@
+DEBUG=True
+SECRET_KEY=@lkf@7pbimd7=c_#5ake7_wxhvjle4s6!5s%kj^017tk&-)e!-
+
+# Database configuration for local development
+DB_NAME=pharmago
+DB_USER=superpharmago
+DB_PASSWORD=pharmagoldenkey
+DB_HOST=127.0.0.1
+DB_PORT=5434
+
+# CORS and security settings
+ALLOWED_HOSTS=localhost,127.0.0.1,0.0.0.0
+CORS_ALLOWED_ORIGINS=http://localhost:3000,http://127.0.0.1:3000
+CSRF_TRUSTED_ORIGINS=http://localhost:3000,http://127.0.0.1:3000
+
+# AWS S3 settings (keep existing)
+AWS_ACCESS_KEY_ID=AKIA5RA7TEOC2ZAU5MG2
+AWS_SECRET_ACCESS_KEY=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
+AWS_STORAGE_BUCKET_NAME=pharmago-user-uploads
+AWS_S3_REGION_NAME=ap-southeast-2
+
+# Redis configuration for local development
+REDIS_URL=redis://localhost:6379/1
+CELERY_BROKER_URL=redis://localhost:6379/0
+CELERY_RESULT_BACKEND=redis://localhost:6379/0
