+++ backend/api/direct/__init__.py
@@ -1,0 +1,7 @@
+"""Direct, read-only endpoints extracted from project urls.
+
+This package intentionally keeps imports inside function bodies in views to
+avoid import-time work and circular dependencies.
+"""
+
+
