+++ backend/api/files/__init__.py
@@ -1,0 +1,6 @@
+"""File/document serving and diagnostics endpoints.
+
+Functions keep imports inside bodies to avoid import-time heaviness.
+"""
+
+
